moved to specific trait, codestyle, phpstan
